Use custom RefreshingCache when for cache that refresh on the background.
Guava cache cannot be made to keep cached entries up to date so that
they will never return outdated data. It only triggers update of a
cached key when requesting the value of the same key. This results to
an outdated value to be used while fetching the new value for the next
caller. Our RefreshingCache updates cached entries on the background and
also does eviction based on actual access instead of when the value was
refreshed. Additionally it applies random time scattering to avoid
huge peridical HTTP bombardments.